node api latest changes

+++ controllers/contactControllers.js
@@ -3,7 +3,7 @@ const Contact = require("../models/contactModel");
 
 // @desc: get method for the contact list  /api/contacts
 const getContacts = asyncHandler(async (req, res) => {
-  const contact = await Contact.find();
+  const contact = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contact);
 });
 
@@ -30,6 +30,7 @@ const postContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   res.status(200).json(contact);
 });
@@ -40,6 +41,13 @@ const putContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
+  }
+  // restrict one loggedin user to update another user contact list
+  if (contact.user_id.toString() != req.user.id) {
+    res.status(403);
+    throw new Error(
+      "user don't have permission to update other user's contact info"
+    );
   }
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
@@ -56,6 +64,13 @@ const deleteContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
+  }
+  // restrict one loggedin user to delete another user contact list
+  if (contact.user_id.toString() != req.user.id) {
+    res.status(403);
+    throw new Error(
+      "user don't have permission to delete other user's contact info"
+    );
   }
   await Contact.deleteOne({ _id: req.params.id });
   res.status(200).json(contact);
